+++ src/PKLorentzVector.cxx
@@ -12,7 +12,7 @@ PKLorentzVector::PKLorentzVector(PKVar x0_,PKVar x1_,PKVar x2_, PKVar x3_)
 	PKVar Magn2_ = X[0]*X[0] - X[1]*X[1] - X[2]*X[2] - X[3]*X[3];
         Magn_ = Magn2_.Sqrt();
 }
-PKLorentzVector::PKLorentzVector(double x0_,double x1_,double x2_,double x3_,double x0err_=0,double x1err_=0,double x2err_=0, double x3err_=0)
+PKLorentzVector::PKLorentzVector(float x0_,float x1_,float x2_,float x3_,float x0err_=0,float x1err_=0,float x2err_=0, float x3err_=0)
 {
 	X[0] = PKVar(x0_,x0err_),X[1] = PKVar(x1_,x1err_),X[2] = PKVar(x2_,x2err_);
 	X[3] = PKVar(x3_,x3err_);
